[FIX] #110 - TodoCard 컴포넌트 분리
